+++ bridge_sdks/python/msgr_signal_bridge/__init__.py
@@ -1,9 +1,13 @@
 """Signal bridge package wiring for Msgr."""
 
 from .client import (
+    HttpResponse,
     LinkingCode,
+    SignalRestClient,
+    SignalServiceError,
     SignalClientProtocol,
     SignalProfile,
+    UrlLibTransport,
     decode_session_blob,
     encode_session_blob,
 )
@@ -12,8 +16,12 @@ from .daemon import SignalBridgeDaemon
 
 __all__ = [
     "LinkingCode",
+    "SignalRestClient",
+    "SignalServiceError",
     "SignalClientProtocol",
     "SignalProfile",
+    "UrlLibTransport",
+    "HttpResponse",
     "decode_session_blob",
     "encode_session_blob",
     "SessionManager",